refactor: fix styles and cards structure, change size of images
